+++ XMonad/Util/PureX.hs
@@ -44,7 +44,7 @@ module XMonad.Util.PureX (
   withWindowSet', withFocii,
   modify'', modifyWindowSet',
   getStack, putStack, peek,
-  focusWindow,
+  focusWindow, focusNth,
   view, greedyView, invisiView,
   shift, curScreen, curWorkspace,
   curTag, curScreenId,
@@ -53,6 +53,7 @@ module XMonad.Util.PureX (
 -- xmonad
 import XMonad
 import qualified XMonad.StackSet as W
+import qualified XMonad.Actions.FocusNth
 
 -- mtl
 import Control.Monad.State
@@ -284,6 +285,10 @@ focusWith focuser = do
 -- | A refresh-tracking version of @W.focusWindow@.
 focusWindow :: XLike m => Window -> m Any
 focusWindow w = focusWith (W.focusWindow w)
+
+-- | A refresh-tracking version of @XMonad.Actions.FocusNth.focusNth@.
+focusNth :: XLike m => Int -> m Any
+focusNth i = focusWith (W.modify' (XMonad.Actions.FocusNth.focusNth' i))
 
 -- }}}